refactor: redux에서 필요없는 함수 삭제
